# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.BetInput:
    properties:
      back:
        description: 后区 2 个号码
        example:
        - 10
        - 12
        items:
          type: integer
        type: array
      front:
        description: 前区 5 个号码
        example:
        - 4
        - 10
        - 12
        - 16
        - 20
        items:
          type: integer
        type: array
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: 'failed to read uploaded image: ...'
        type: string
    type: object
...
  handler.ParseTicketResponse:
    properties:
      bets:
        description: 规范化后的号码
        items:
          $ref: '#/definitions/lottery.SuperLotto'
        type: array
      issue:
        description: 识别出的期号
        type: string
      total_bets:
        description: 注数
        type: integer
    type: object
...
  handler.ScanResponse:
    properties:
      raw:
        description: QR 码解码出的原始文本
        type: string
      ticket:
        allOf:
        - $ref: '#/definitions/lottery.SuperLotto'
        description: 识别出的号码
    type: object
  handler.TicketInput:
    properties:
      bets:
        description: 票面号码；与 text 二选一
        items:
          $ref: '#/definitions/handler.BetInput'
        type: array
      issue:
        description: 期号（可选）
        example: "26102"
        type: string
      text:
        description: 整票文本，一行一注；bets 为空时使用
        type: string
    type: object
...
  handler.VerifyTicketRequest:
    properties:
      bets:
        description: 票面号码；与 text 二选一
        items:
          $ref: '#/definitions/handler.BetInput'
        type: array
      issue:
        description: 期号（可选）
        example: "26102"
        type: string
      text:
        description: 整票文本，一行一注；bets 为空时使用
        type: string
      winning_back:
        description: 开奖后区（可选）
        example: 06,07
        type: string
      winning_front:
        description: 开奖前区（可选）
        example: 01,03,07,27,28
        type: string
    type: object
  lottery.BetResult:
    properties:
      amount:
        description: 按奖池档位算出的单注奖金
        type: string
      back:
        description: 后区号码
        items:
          type: integer
        type: array
      front:
        description: 前区号码
        items:
          type: integer
        type: array
      index:
        description: 注序号（从 1 开始）
        type: integer
      matched_back:
        description: 命中后区个数
        type: integer
      matched_front:
        description: 命中前区个数
        type: integer
      prize:
        $ref: '#/definitions/lottery.Prize'
      won:
        description: 该注是否中奖
        type: boolean
    type: object
  lottery.Prize:
    properties:
      amount_above_8yi:
        description: 奖池 8 亿以上单注奖金
        type: string
      amount_below_8yi:
        description: 奖池 8 亿以下单注奖金
        type: string
      condition:
        description: 中奖条件，如"5+2"
        type: string
      tier:
        description: 奖级（一等奖=1 … 七等奖=7）
        type: integer
      tier_name:
        description: 奖级名称，如"一等奖"
        type: string
    type: object
  lottery.SuperLotto:
    properties:
      back:
        description: 后区号码
        example:
        - 1
        - 8
        items:
          type: integer
        type: array
      front:
        description: 前区号码
        example:
        - 5
        - 12
        - 18
        - 23
        - 35
        items:
          type: integer
        type: array
    type: object
  lottery.TicketVerifyResult:
    properties:
      bets:
        description: 逐注结果
        items:
          $ref: '#/definitions/lottery.BetResult'
        type: array
      issue:
        description: 期号
        type: string
      total_bets:
        description: 总注数
        type: integer
      total_won:
        description: 是否至少中一注
        type: boolean
      winning:
        allOf:
        - $ref: '#/definitions/lottery.WinningNumbers'
        description: 使用的开奖号码
      winning_bets:
        description: 中奖注数
        type: integer
    type: object
  lottery.VerifyResult:
    properties:
      matched_back:
        description: 命中后区个数
        type: integer
      matched_front:
        description: 命中前区个数
        type: integer
      prize:
        allOf:
        - $ref: '#/definitions/lottery.Prize'
        description: 中奖信息（未中奖时为 null）
      won:
        description: 是否中奖
        type: boolean
    type: object
  lottery.WinningNumbers:
    properties:
      back:
        description: 后区开奖号码
        example:
        - 6
        - 7
        items:
          type: integer
        type: array
      draw_time:
        description: 开奖日期
        example: "2026-09-07"
        type: string
      front:
        description: 前区开奖号码
        example:
        - 1
        - 3
        - 7
        - 27
        - 28
        items:
          type: integer
        type: array
      issue:
        description: 期号
        example: "26102"
        type: string
      pool_above_8yi:
        description: 奖池是否达到 8 亿
        type: boolean
      pool_balance:
        description: 开奖后奖池（元）
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: wechatapp-web
    url: http://www.swagger.io/support
  description: |-
    Image background replacement API. Uploads an image, removes its
    background via the remove.bg service, then optionally composites
    the cutout onto a new background (solid color or another image).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: wechatapp-web API
  version: 1.0.0
paths:
...
  /lottery/parse-ticket:
    post:
      consumes:
      - application/json
      description: |-
        Validates client-supplied ticket numbers (from manual input or
        on-device OCR) and returns them normalized, so the app can show
        a confirmation screen before calling /lottery/verify-ticket.
      parameters:
      - description: Ticket numbers as bets or text
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.TicketInput'
      produces:
      - application/json
      responses:
        "200":
          description: Normalized issue and bets
          schema:
            $ref: '#/definitions/handler.ParseTicketResponse'
        "400":
          description: Missing input or invalid numbers (message names the offending
            bet/line)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Parse and normalize 大乐透 ticket numbers
      tags:
      - lottery
  /lottery/scan:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Decodes a QR-code image and extracts the 大乐透 numbers printed
        by the user (5 front numbers + 2 back numbers). The QR payload
        must contain 7 numbers, e.g. "05,12,18,23,35,01,08".
      parameters:
      - description: QR code image containing the ticket numbers
        in: formData
        name: qrcode
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Decoded raw text and recognized numbers
          schema:
            $ref: '#/definitions/handler.ScanResponse'
        "400":
          description: Missing qrcode, unreadable image, or unparseable numbers
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Image decodes but contains no valid QR code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Recognize 大乐透 numbers from a QR code
      tags:
      - lottery
  /lottery/verify:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Scans the QR code to get the user's printed numbers, then
        compares them with the given winning numbers and returns the
        prize tier if the ticket wins.
      parameters:
      - description: QR code image with the user's printed numbers
        in: formData
        name: qrcode
        required: true
        type: file
      - description: Winning front-area numbers, e.g. 05,12,18,23,35
        in: formData
        name: winning_front
        required: true
        type: string
      - description: Winning back-area numbers, e.g. 01,08
        in: formData
        name: winning_back
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Match counts and prize (won=false if no prize)
          schema:
            $ref: '#/definitions/lottery.VerifyResult'
        "400":
          description: Missing fields or invalid numbers
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Image decodes but contains no valid QR code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Verify a 大乐透 ticket against winning numbers
      tags:
      - lottery
  /lottery/verify-ticket:
    post:
      consumes:
      - application/json
      description: |-
        Verifies a printed 大乐透 ticket that may contain several bets
        (单式票通常 5 注). Numbers may be sent as structured bets or as
        multi-line text. Winning numbers can be supplied directly; if
        omitted and an issue is given, they are fetched from the
        official China Sports Lottery draw-history API.
      parameters:
      - description: Ticket bets and optional winning numbers
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.VerifyTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Per-bet match counts, prize and totals
          schema:
            $ref: '#/definitions/lottery.TicketVerifyResult'
        "400":
          description: Invalid numbers or missing winning numbers/issue
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "502":
          description: Failed to fetch official winning numbers
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Verify a multi-bet 大乐透 ticket
      tags:
      - lottery
  /lottery/winning/{issue}:
    get:
      description: |-
        Fetches the official draw result for the given issue from the
        China Sports Lottery API (cached in memory).
      parameters:
      - description: Issue number, e.g. 26102
        in: path
        name: issue
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Official draw result
          schema:
            $ref: '#/definitions/lottery.WinningNumbers'
        "404":
          description: Issue not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "502":
          description: Upstream API error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get official 大乐透 winning numbers
      tags:
      - lottery
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT issued by POST /auth/login, sent as "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"